Converted MUIIcon to Icon
