+++ apps/nextjs-website/src/components/Header.tsx
@@ -3,7 +3,7 @@ import { Header as HeaderEC } from '@pagopa/pagopa-editorial-components/dist/com
 import { HeaderProps } from '@pagopa/pagopa-editorial-components/dist/components/Header/Header';
 import { Stack } from '@mui/material';
 import { usePathname } from 'next/navigation';
-import MUIIcon from './MUIIcon';
+import MUIIcon from './Icon';
 import { HeaderWithNavigation } from '@/lib/header';
 
 const makeHeaderProps = (
