--- apps/nextjs-website/src/lib/fetch/types/InternalIcons.ts
@@ -1,7 +1,0 @@
-import * as t from 'io-ts';
-
-export const InternalSocialIconCodec = t.keyof({
-  Medium: null,
-});
-
-export type InternalSocialIcon = t.TypeOf<typeof InternalSocialIconCodec>;
