+++ apps/nextjs-website/src/lib/fetch/types/icons/FeatureItemIcon.ts
