+++ apps/nextjs-website/src/lib/fetch/types/icons/HowToStepIcon.ts
