+++ apps/nextjs-website/src/lib/fetch/types/icons/SocialIcon.ts
@@ -1,0 +1,9 @@
+import * as t from 'io-ts';
+
+export const SocialIconCodec = t.keyof({
+  Instagram: null,
+  LinkedIn: null,
+  Facebook: null,
+  Twitter: null,
+  Medium: null,
+});
