+++ apps/nextjs-website/src/lib/fetch/types/icons/StripeLinkIcon.ts
